refs 124384: move external user group files to features/external-user-group
