+++ apps/app/src/components/Admin/UserGroup/ExternalUserGroup/LdapGroupSyncSettingsForm.tsx
@@ -6,7 +6,7 @@ import { useTranslation } from 'react-i18next';
 
 import { apiv3Put } from '~/client/util/apiv3-client';
 import { toastError, toastSuccess } from '~/client/util/toastr';
-import { LdapGroupMembershipAttributeType, LdapGroupSyncSettings } from '~/interfaces/external-user-group';
+import { LdapGroupMembershipAttributeType, LdapGroupSyncSettings } from '~/features/external-user-group/interfaces/external-user-group';
 import { useSWRxLdapGroupSyncSettings } from '~/stores/external-user-group';
 
 export const LdapGroupSyncSettingsForm: FC = () => {
